Move api tests to own file
